deepseek-coder/requirements.txt
pytorch
accelerate
peft
bitsandbytes
transformers
trl
wandb
huggingface_hub